canvas_quiz_uploader/requirements.txt
streamlit==1.37.1
requests==2.32.3
python-docx==1.1.2
